+++ cmd/deploy-with-helm/main.go
@@ -1,19 +1,19 @@
 package main
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"flag"
 	"fmt"
 	"io/fs"
 	"io/ioutil"
 	"log"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"strings"
 
-	"github.com/google/shlex"
 	"github.com/opendevstack/pipeline/internal/command"
 	"github.com/opendevstack/pipeline/internal/directory"
 	"github.com/opendevstack/pipeline/internal/file"
@@ -24,7 +24,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
-	"sigs.k8s.io/yaml"
 )
 
 const (
@@ -36,27 +35,12 @@ const (
 	ageKeyFilePath = "./key.txt"
 )
 
-// confusingHelmDiffMessage is the message Helm prints when helm-diff is
-// configured to exit with a non-zero exit code when drift is detected,
-// provided helm-secrets is also installed but no secrets are decrypted.
-const confusingHelmDiffMessage = `Error: identified at least one change, exiting with non-zero exit code (detailed-exitcode parameter enabled)
-Error: plugin "diff" exited with error
-Error: plugin "secrets" exited with error`
-
-// confusingHelmDiffDecryptionMessage is like confusingHelmDiffMessage but occurs
-// when helm-secrets is required to decrypt secrets to perform the diff.
-const confusingHelmDiffDecryptionMessage = `Error: identified at least one change, exiting with non-zero exit code (detailed-exitcode parameter enabled)
-Error: plugin "diff" exited with error
-
-/usr/local/helm/plugins/helm-secrets/scripts/commands/helm.sh: line 34: xargs: command not found
-Error: plugin "secrets" exited with error`
-
 type options struct {
 	// Location of Helm chart directory.
 	chartDir string
 	// Name of Helm release.
 	releaseName string
-	// Flags to pass to `helm diff upgrade` (in addition to default ones).
+	// Flags to pass to `helm diff upgrade` (in addition to default ones and upgrade flags).
 	diffFlags string
 	// Flags to pass to `helm upgrade`.
 	upgradeFlags string
@@ -76,7 +60,7 @@ func main() {
 	opts := options{}
 	flag.StringVar(&opts.chartDir, "chart-dir", "", "Chart dir")
 	flag.StringVar(&opts.releaseName, "release-name", "", "Name of Helm release")
-	flag.StringVar(&opts.diffFlags, "diff-flags", "", "Flags to pass to `helm diff upgrade` (in addition to default ones)")
+	flag.StringVar(&opts.diffFlags, "diff-flags", "", "Flags to pass to `helm diff upgrade` (in addition to default ones and upgrade flags)")
 	flag.StringVar(&opts.upgradeFlags, "upgrade-flags", "", "Flags to pass to `helm upgrade`")
 	flag.StringVar(&opts.ageKeySecret, "age-key-secret", "", "Name of the secret containing the age key to use for helm-secrets")
 	flag.StringVar(&opts.ageKeySecretField, "age-key-secret-field", "key.txt", "Name of the field in the secret holding the age private key")
@@ -116,11 +100,11 @@ func main() {
 	// read ods.y(a)ml
 	odsConfig, err := config.ReadFromDir(checkoutDir)
 	if err != nil {
-		log.Fatal(fmt.Sprintf("err during ods config reading: %s", err))
+		log.Fatalf("err during ods config reading: %s", err)
 	}
 	targetConfig, err := odsConfig.Environment(ctxt.Environment)
 	if err != nil {
-		log.Fatal(fmt.Sprintf("err during namespace extraction: %s", err))
+		log.Fatalf("err during namespace extraction: %s", err)
 	}
 
 	releaseNamespace := targetConfig.Namespace
@@ -371,32 +355,25 @@ func main() {
 	}
 
 	fmt.Printf("Diffing Helm release against %s...\n", helmArchive)
-	helmDiffArgs := []string{
-		"--namespace=" + releaseNamespace,
-		"secrets",
-		"diff",
-		"upgrade",
-		"--detailed-exitcode",
-		"--no-color",
-	}
-	helmDiffFlags, err := shlex.Split(opts.diffFlags)
+	helmDiffArgs, err := assembleHelmDiffArgs(
+		releaseNamespace, releaseName, helmArchive,
+		opts, valuesFiles, cliValues,
+	)
 	if err != nil {
-		log.Fatalf("could not parse diff flags (%s): %s", opts.diffFlags, err)
+		log.Fatal(err)
 	}
-	helmDiffArgs = append(helmDiffArgs, helmDiffFlags...)
-	for _, vf := range valuesFiles {
-		helmDiffArgs = append(helmDiffArgs, fmt.Sprintf("--values=%s", vf))
-	}
-	helmDiffArgs = append(helmDiffArgs, cliValues...)
-	helmDiffArgs = append(helmDiffArgs, releaseName, helmArchive)
 	stdout, stderr, err = runHelmCmd(helmDiffArgs, targetConfig, opts.debug)
-	if err == nil {
-		fmt.Println("no diff ...")
+	if err == nil { // exit code 0 returned
+		fmt.Println("No diff detected, skipping helm upgrade.")
 		os.Exit(0)
+	}
+	var ee *exec.ExitError
+	if errors.As(err, &ee) && ee.ExitCode() != diffExitCode { // exit code 1 returned
+		log.Fatalf("%s\n%s", err, string(ee.Stderr))
 	}
 	// Replace confusing stderr messages while still printing stderr in general
 	// to surface any other issues that might be logged there.
-	diffStderr := replaceConfusingHelmLogMessages(stderr)
+	diffStderr := cleanHelmDiffOutput(stderr)
 	fmt.Println(string(stdout))
 	fmt.Println(string(diffStderr))
 	err = writeDeploymentArtifact(stdout, "diff", opts.chartDir, targetConfig.Name)
@@ -405,21 +382,13 @@ func main() {
 	}
 
 	fmt.Printf("Upgrading Helm release to %s...\n", helmArchive)
-	helmUpgradeArgs := []string{
-		"--namespace=" + releaseNamespace,
-		"secrets",
-		"upgrade",
-	}
-	helmUpgradeFlags, err := shlex.Split(opts.upgradeFlags)
+	helmUpgradeArgs, err := assembleHelmUpgradeArgs(
+		releaseNamespace, releaseName, helmArchive,
+		opts, valuesFiles, cliValues,
+	)
 	if err != nil {
-		log.Fatalf("could not parse upgrade flags (%s): %s", opts.upgradeFlags, err)
+		log.Fatal(err)
 	}
-	helmUpgradeArgs = append(helmUpgradeArgs, helmUpgradeFlags...)
-	for _, vf := range valuesFiles {
-		helmUpgradeArgs = append(helmUpgradeArgs, fmt.Sprintf("--values=%s", vf))
-	}
-	helmUpgradeArgs = append(helmUpgradeArgs, cliValues...)
-	helmUpgradeArgs = append(helmUpgradeArgs, releaseName, helmArchive)
 	stdout, stderr, err = runHelmCmd(helmUpgradeArgs, targetConfig, opts.debug)
 	if err != nil {
 		fmt.Println(string(stderr))
@@ -430,51 +399,6 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-type helmChart struct {
-	Name    string `json:"name"`
-	Version string `json:"version"`
-}
-
-func getHelmChart(filename string) (*helmChart, error) {
-	y, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, fmt.Errorf("could not read chart: %w", err)
-	}
-
-	var hc *helmChart
-	err = yaml.Unmarshal(y, &hc)
-	if err != nil {
-		return nil, fmt.Errorf("could not unmarshal: %w", err)
-	}
-	return hc, nil
-}
-
-func getChartVersion(contextVersion string, hc *helmChart) string {
-	if len(contextVersion) > 0 && contextVersion != pipelinectxt.WIP {
-		return contextVersion
-	}
-	return hc.Version
-}
-
-// replaceConfusingHelmLogMessages replaces confusing stderr messages with
-// easier to understand ones. The tests in ods-deploy-helm_test.go ensure this
-// still works when Helm is upgraded.
-func replaceConfusingHelmLogMessages(logs []byte) []byte {
-	cleaned := bytes.Replace(
-		logs,
-		[]byte(confusingHelmDiffMessage),
-		[]byte("plugin \"diff\" identified at least one change"),
-		1,
-	)
-	cleaned = bytes.Replace(
-		cleaned,
-		[]byte(confusingHelmDiffDecryptionMessage),
-		[]byte("plugin \"diff\" identified at least one change"),
-		1,
-	)
-	return cleaned
 }
 
 func artifactFilename(filename, chartDir, targetEnv string) string {
@@ -507,33 +431,6 @@ func storeAgeKey(secret *corev1.Secret, ageKeySecretField string) (errBytes []by
 	return errBytes, err
 }
 
-func runHelmCmd(args []string, targetConfig *config.Environment, debug bool) (outBytes, errBytes []byte, err error) {
-	if debug {
-		args = append([]string{"--debug"}, args...)
-	}
-	printableArgs := args
-	if targetConfig.APIServer != "" {
-		printableArgs = append(
-			[]string{
-				fmt.Sprintf("--kube-apiserver=%s", targetConfig.APIServer),
-				"--kube-token=***",
-			},
-			args...,
-		)
-		args = append(
-			[]string{
-				fmt.Sprintf("--kube-apiserver=%s", targetConfig.APIServer),
-				fmt.Sprintf("--kube-token=%s", targetConfig.APIToken),
-			},
-			args...,
-		)
-	}
-	fmt.Println(helmBin, strings.Join(printableArgs, " "))
-
-	var extraEnvs = []string{fmt.Sprintf("SOPS_AGE_KEY_FILE=%s", ageKeyFilePath)}
-	return command.RunWithExtraEnvs(helmBin, args, extraEnvs)
-}
-
 func tokenFromSecret(clientset *kubernetes.Clientset, namespace, name string) (string, error) {
 	secret, err := clientset.CoreV1().Secrets(namespace).Get(context.TODO(), name, metav1.GetOptions{})
 	if err != nil {
@@ -548,33 +445,6 @@ func getTrimmedFileContent(filename string) (string, error) {
 		return "", err
 	}
 	return strings.TrimSpace(string(content)), nil
-}
-
-func packageHelmChart(chartDir, ctxtVersion, gitCommitSHA string, debug bool) (string, error) {
-	hc, err := getHelmChart(filepath.Join(chartDir, "Chart.yaml"))
-	if err != nil {
-		return "", fmt.Errorf("could not read chart: %w", err)
-	}
-	chartVersion := getChartVersion(ctxtVersion, hc)
-	packageVersion := fmt.Sprintf("%s+%s", chartVersion, gitCommitSHA)
-	helmPackageArgs := []string{
-		"package",
-		fmt.Sprintf("--app-version=%s", gitCommitSHA),
-		fmt.Sprintf("--version=%s", packageVersion),
-	}
-	if debug {
-		helmPackageArgs = append(helmPackageArgs, "--debug")
-	}
-	stdout, stderr, err := command.Run(helmBin, append(helmPackageArgs, chartDir))
-	if err != nil {
-		return "", fmt.Errorf(
-			"could not package chart %s. stderr: %s, err: %s", chartDir, string(stderr), err,
-		)
-	}
-	fmt.Println(string(stdout))
-
-	helmArchive := fmt.Sprintf("%s-%s.tgz", hc.Name, packageVersion)
-	return helmArchive, nil
 }
 
 func collectImageDigests(imageDigestsDir string) ([]string, error) {